chore: replace image links with the value of `currentSrc` at runtime
This will take the user to a full-screen version of the image they are
already looking at.

I originally wanted the link to take the user to the original, but given
the new 1 MiB limit (f39cf30), the version of the image the user opens
might actually be of lesser quality than the one they can already see.

Since the processor is aware of the current commit hash, I might just
replace this link with one going to GitHub in the future. That would
make the site less standalone, though.