chore: de-duplicate assets and make 100% of `public` generated
The structure of the public directory is left over from when I was
considering committing the whole public directory with every commit,
before deciding this was impractical.

Having just one directory where web assets are stored, while it's
confusing that some are used to generate pages then discarded, and some
are now copied; I think it's simpler to understand at a glance.